humanity/requirements.txt
web3==6.15.1
pyyaml==6.0.1 